Small changes on Viewmodels
Implemented viewmodel on the other textfields with the UserTest data class, implemented functions to clear textfield. To Do: a way to use only one method to update textfield. Problem incurred: difficulties to do this method because cant pass a parametres to changes only one textfield, they change all in once