Changes on the viewmodel methods
Found a new way to implement correct data flow with viewmodel using only one method to update the state of user data, with the creation of a sealed class used for the registration events. Also some test composable to see if it works. (This commit was made to show these new implementations)