Bid Input Changes
Divided the different bid for AuctionType.
Implemente a new class with methods for the correct input for decimal value for bid input, with limitation for the input to write (methods thanks to a formatted text, the value will be always support for the to.Float() method which need to use for the correct viewModel Flow).
ToDo: Find a way to not use a mutableStateOf for the textfield that requires non-String values.