ViewModel Changes
Refactor name for a MainViewModel.
Added ViewModel for the RegistrationView.
New methos on this viewModel with the help of a EventClass for cleaner implementation.
Implemented a validation with error checks and msg.
New RegistrationStateClass for the State of the registration view (Tests to see if we can use always the user data class and only a new error class for that form).
ToDo: ContactInfo Composable and CreditcardField to implement in a correct way with viewModels (Have to see if is better to use the RegistrationState or the UserData class as a StateFlow and a Error State class)