SharedViewModel
Tested sharedViewModel, with a test state (sharedState) that passes correctly from AuctionView to MakeABid with a test udpate methods to see if the changes correctly passes through these views.
In the SetupNavGraph these 2 views are nested thanks to the navigation fun, and this is implemented that when in this nested graph,  the views uses a SharedViewModel, and on quitting this nest, the SharedViewModel clears.