Nav Shared Params
Modified the navigation from home to a specific auction thanks to shared params on the navigation arguments.
ToDo: the fetchAuction method on the shared viewmodel has to take the auctionId passed in the navigation argument and update the auctionState on the Init of the screen and then show the auctionState on the View.