Changes to createAuction Uris handling
The ViewModel now handles strings, so the composable passes strings to the viewmodel instead of Uris when uploading to memory. This enables the ViewModel to create a File instance for uploading the image to an S3 bucket and retrieving its public URL. The public URL is then added to the list of items to be stored in the backend database. Meanwhile, ImageItem can handle the Uri as a string without issues.